note:  I just returned working on this project.
I was taking a time to rest and work on other non programming related stuff, like finishing my LinkedIn profile info, witch is necessary for this project by the way. I also needed time to get drunk and fuck shit up...that's right, I'm tired of putting so much effort on this field, making commits for 33 days straight, writing english everyday even though its not my main language, trying to learn something new just not to stay behind, I know that's the right path but we need to stop and just enjoy being alive, people are dying everyday, it can be your mom, dad, friend ou anyone really, so why waste so much time? Life is just wonderful the way it is right now, there's no point in giving your life for a future that is not in the present... So yeah I'm back